[ refactor ] SNS Share 관련 코드 리팩토링
