+++ components/button/snsBox.tsx
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
 
-interface SnsBoxProps {
+interface SNSBoxProps {
   children: React.ReactNode;
   onClick?: () => void;
 }
 
-export default function SnsBox(props: SnsBoxProps) {
+export default function SNSBox(props: SNSBoxProps) {
   const { children, onClick } = props;
 
   return <Styled.Box onClick={onClick}>{children}</Styled.Box>;
@@ -13,7 +13,7 @@ export default function SnsBox(props: SnsBoxProps) {
 
 const Styled = {
   Box: styled.div`
-    cursor: pointer;
     margin: 0 0.5rem 0;
+    cursor: pointer;
   `,
 };
